+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 - Added Query.first()
 - Added Query.firstOrDefault()
 - Added Query.toList(converter)
+- Added Cursor.toList(converter)
 
 #### v0.1.0 (2019-11-22)
 - Initial release